+++ src/layout/DesktopNavBar.tsx
@@ -6,8 +6,6 @@ import favourites from '../assets/nav_bar/favourites.svg'
 import favouritesActive from '../assets/nav_bar/favourites_filled.svg'
 import profile from '../assets/nav_bar/profile.svg'
 import profileActive from '../assets/nav_bar/profile_filled.svg'
-import track from "../assets/nav_bar/track.svg"
-import trackActive from "../assets/nav_bar/track_filled.svg"
 import { Link, useLocation } from 'react-router-dom'
 
 const DesktopNavBar = () => {