+++ src/pages/BatchImport/index.tsx
@@ -6,11 +6,7 @@ import InputImportContacts from '../../components/InputImportContacts';
 function BatchImportModal() {
   const {  dismissImportContacts, _showImportContacts } = useContext(appContext);
   const [step, setStep] = useState(1);
-  const [isLoading, setIsLoading] = useState(false);
   const [_length, setLength] = useState(null);
-  const [_current, setCurrent] = useState(null);
-  const [_passed, setPassed] = useState(null);
-  const [_failed, setFailed] = useState(null);
   
 
 
@@ -31,28 +27,27 @@ function BatchImportModal() {
               {step === 1 && (
                 <>
                   <h5 className="font-bold text-lg mt-3 mb-5">Batch Import</h5>
-                  <p className="mb-5">Attach your exported Contacts JSON file below</p>
-                  <InputImportContacts setStep={setStep} setLength={setLength} setCurrent={setCurrent} setFailed={setFailed} setPassed={setPassed} />
+                  <p className="mb-5">Attach your exported contacts txt file below</p>
+                  <InputImportContacts setStep={setStep} setLength={setLength} />
 
-                  {!isLoading && (
-                    <div className="pt-5 pb-3 text-center">
-                      <div onClick={dismiss} className="cursor-pointer text-custom-grey font-medium link">
-                        Cancel
-                      </div>
+                  
+                  <div className="pt-5 pb-3 text-center">
+                    <div onClick={dismiss} className="cursor-pointer text-custom-grey font-medium link">
+                      Cancel
                     </div>
-                  )}
+                  </div>
                 </>
               )}
               {step === 2 && (
                 <>
                   <h5 className="text-lg mt-6 mb-10">
-                    <strong className="capitalize">Importing {_current}/{_length} contacts...</strong>
+                    <strong className="capitalize">Importing contacts...</strong>
                   </h5>
                   <div className="w-32 h-32 flex items-center justify-center text-xs text-left mx-auto mb-16">
                     <img src={onboardingIcon} width={135} alt="Onboarding icon" />
                   </div>
                   <div className="mb-1">
-                    <button disabled={true} className="cursor-pointer text-white w-full text-base font-bold py-3 rounded rounded-xl bg-custom-purple bg-custom-purple--disabled">
+                    <button disabled={true} className="cursor-pointer text-white w-full text-base font-bold py-3 rounded-xl bg-custom-purple bg-custom-purple--disabled">
                       Close
                     </button>
                   </div>
@@ -60,30 +55,17 @@ function BatchImportModal() {
               )}
               {step === 3 && (
                 <>
-                  {_passed !== null && _passed === 0 && 
-                    <h5 className="text-lg mt-6 mb-10">
-                      No contacts were imported
-                    </h5>
-                  }
-                  {_passed !== null && _passed > 0 && 
-                    <h5 className={`text-lg mt-6 ${_failed !== null && _failed === 0 ? 'mb-10' : 'mb-5' }`}>
-                      {_passed} contacts were imported
-                    </h5>
-                  }
-                  
-                  
-                  {_failed !== null && _failed > 0 &&                  
-                  <h5 className="text-lg mb-10">
-                    <strong>{_failed}</strong> failed to be imported
+                  <h5 className="text-lg mt-6 mb-10">
+                     Imported {_length} contacts!
                   </h5>
-                  }
+                  
                   <div className="w-32 h-32 rounded-full flex items-center justify-center text-xs text-left mx-auto mb-16" style={{ background: '#7A17F9' }}>
                     <svg width="49" height="36" viewBox="0 0 49 36" fill="none" xmlns="http://www.w3.org/2000/svg">
                       <path d="M17.2468 35.1763L0.804199 18.7337L3.96058 15.5773L17.2468 28.8635L45.434 0.67627L48.5904 3.83265L17.2468 35.1763Z" fill="white" />
                     </svg>
                   </div>
                   <div className="mb-1">
-                    <button onClick={dismiss} className="cursor-pointer text-white w-full text-base font-bold py-3 rounded rounded-xl bg-custom-purple">
+                    <button onClick={dismiss} className="cursor-pointer text-white w-full text-base font-bold py-3 rounded-xl bg-custom-purple">
                       Close
                     </button>
                   </div>