+++ src/utilities/exportToJson/index.ts
@@ -1,15 +1,31 @@
-export const exportToJson = (data: any, fileName: string) => {
+export const exportToJson = (data: string, fileName: string) => {
 
   try {
-    const blob = new Blob([JSON.stringify(data)], {type: 'application/json'});
+    const blob = new Blob([data], {type: 'application/text'});
     const url = URL.createObjectURL(blob);
 
+    if (
+      window.navigator.userAgent.includes("Minima Browser")
+    ) {
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
+      return Android.fileDownload(
+        (window as any).MDS.minidappuid,
+        fileName
+      );
+    }
+
+    
     const a = document.createElement('a');
     a.href = url;
-    a.download = `${fileName}.json`;
+    a.download = `${fileName}.txt`;
     a.click();
-    document.body.removeChild(a);
+    
+    if (document.body.contains(a)) {
+      document.body.removeChild(a);
+    }
     URL.revokeObjectURL(url);
+    
 
     return true;
   } catch (error) {